requirements.txt
asgiref==3.6.0
autopep8==2.0.2
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
cryptography==40.0.0
defusedxml==0.7.1
Django==4.1.7
django-allauth==0.53.1
django-cors-headers==3.14.0
django-rest-auth==0.9.5
django-rest-auth-forked==0.9.5
django-suit==0.2.28
djangorestframework==3.14.0
idna==3.4
oauthlib==3.2.2
Pillow==9.4.0
psycopg2-binary==2.9.5
pycodestyle==2.10.0
pycparser==2.21
PyJWT==2.6.0
python3-openid==3.2.0
pytz==2022.7.1
requests==2.28.2
requests-oauthlib==1.3.1
six==1.16.0
social-auth-app-django==5.1.0
social-auth-core==4.4.0
sqlparse==0.4.3
tomli==2.0.1
urllib3==1.26.15
